Added newline char to merge script
The merge script was not adding a newline char before the "---" separator when it was merging the .yaml documents. e.g. instead of:

  selector:
    app: httpbin
---

I was getting:

  selector:
    app: httpbin---

which caused problems in the deployment job in the CircleCI pipeline. This change in the merge.sh script fixed this problem, at least in my case.

Note: I had to enclose the characters with a single quote ( ' ) instead of double quotes ( " ) for this to work.